Fixed cmdlet naming in eid-provider-ps
